printk: Add consoles to a virtual "console" bus
This patch embeds a device struct in the console struct, and registers
them on a "console" bus so we can expose attributes in sysfs.

Currently, most drivers declare static console structs, and that is
incompatible with the dev refcount model. So we end up needing to patch
all of the console drivers to:

	1. Dynamically allocate the console struct using a new helper
	2. Handle the allocation in (1) possibly failing
	3. Dispose of (1) with put_device()

Early console structures must still be static, since they're required
before we're able to allocate memory. The least ugly way I can come up
with to handle this is an "is_static" flag in the structure which makes
the gets and puts NOPs, and is checked in ->release() to catch mistakes.

Signed-off-by: Calvin Owens <calvinowens@fb.com>