mm, swap: Implement swap cluster lookup using xarray
When swap extension happens, we may need to add more clusters to the
existing pool. This is non-trivial in the current case because clusters
are contained within one struct swap_cluster_info array, which cannot be
realloc()ed because it contains a spinlock_t, which has tendrils outside
of its immediate type (eg. in lockdep).

There are a couple of possible approaches:

1. Turn the array of clusters into multi-dimensional, `struct
   swap_cluster_info**` array. This can work, but eventually still
   requires us to implement yet another source of allocation management.
2. Use some other datastructure. XArray seems a reasonable fit, doesn't
   put a dent in typical swap latencies or take too much space for
   multi-gigabyte swap files, and is well supported.