mm: swap: Move to using nr_extents
No functional changes.

This code confused me a bit, even as somebody working on a patchset
heavily involving it. This mostly centered around all the special
behaviours of ret, including (but not only) being the number of extents
returned and the implications therein.

This patch makes some minor cosmetic changes that would have helped me
when I was getting started with the patchset.