printk: Userspace format enumeration support
We have a number of systems industry-wide that have a subset of their
functionality that works as follows:

1. Receive a message from local kmsg, serial console, or netconsole;
2. Apply a set of rules to classify the message;
3. Do something based on this classification (like scheduling a
   remediation for the machine), rinse, and repeat.

As a couple of examples of places we have this implemented just inside
Facebook, although this isn't a Facebook-specific problem, we have this
inside our netconsole processing (for alarm classification), and as part
of our machine health checking. We use these messages to determine
fairly important metrics around production health, and it's important
that we get them right.

While for some kinds of issues we have counters, tracepoints, or metrics
with a stable interface which can reliably indicate the issue, in order
to react to production issues quickly we need to work with the interface
which most kernel developers use when developing: printk. Most
production issues come from unexpected phenomena, and as such usually
the code in question doesn't have easily usable tracepoints or other
counters available for the specific problem being mitigated. We have a
number of lines of monitoring defence against problems in production
(host metrics, process metrics, service metrics, etc), and where it's
not possible to monitor at another level, this kind of pragmatic
netconsole monitoring is essential.

As you'd expect, monitoring using printk is rather brittle for a number
of reasons -- most notably that the message might disappear entirely in
a new version of the kernel, or that the message may change in some way
that the regex or other classification methods start to silently fail.

One factor that makes this even harder is that, under normal operation,
many of these messages are never expected to be hit. For example, there
may be some rare hardware bug which you want to detect if it was to ever
happen again, but its recurrence is not likely or anticipated. This
precludes using something like checking for this printk being printed
fleetwide recently to determine whether the message in question is still
present or not.

This class of issue has happened on a number of occasions, causing
pathologically unhealthy machines with hardware issues to remain in
production for longer than ideal. As a recent example, some monitoring
around blk_update_request fell out of date and caused semi-broken
machines to remain in production for longer than would be desirable.

Searching through the codebase to find the message is also extremely
fragile, because many of the messages are further constructed beyond
their callsite (eg. btrfs_printk and other module-specific wrappers,
each with their own functionality). Even if they aren't, guessing the
format and formulation of the underlying message based on the aesthetics
of the message emitted is not a recipe for success at scale, and our
previous issues with fleetwide machine health checking demonstrate as
much.

This patch provides a solution to the issue of silently changed or
deleted printks: we record all printk format strings known at compile
time into a new .printk_fmts section, both in vmlinux and modules. At
runtime, this can then be iterated by looking at /proc/printk_formats,
which emits the same format as `printk` itself, which we already export
elsewhere (for example, in netconsole):

    # Format: <facility><level><format>\0
    $ perl -p -e 's/\n/\\n/g;s/\0/\n/g' /proc/printk_formats | shuf -n 5
    6                       (parent %llu not aligned to sectorsize %u)\n
    6printk: log_buf_len min size: %d bytes\n
    6pci %04x:%02x:%02x.%d: not ready after %dms; waiting\n
    3SLUB: Unable to add boot slab alias %s to sysfs\n
    4hpet: Global config: Unknown bits %#x\n

This mitigates the majority of cases where we have a highly-specific
printk which we want to match on, as we can now enumerate and check
whether the format changed or the printk callsite disappeared entirely
in userspace. This allows us to catch changes to printks we monitor
earlier and decide what to do about it before it becomes problematic.

There are no changes to the generated assembly for printk or callers at
runtime.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Petr Mladek <pmladek@suse.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: John Ogness <john.ogness@linutronix.de>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: linux-kernel@vger.kernel.org
Cc: kernel-team@fb.com