printk: Userspace format enumeration support
We have a number of systems industry-wide that have a subset of their
functionality that works as follows:

1. Receive a message from local kmsg, serial console, or netconsole;
2. Apply a set of rules to classify the message;
3. Do something based on this classification (like scheduling a
   remediation for the machine), rinse, and repeat.

As a couple of examples of places we have this implemented just inside
Facebook, although this isn't a Facebook-specific problem, we have this
inside our netconsole processing (for alarm classification), and as part
of our machine health checking. We use these messages to determine
fairly important metrics around production health, and it's important
that we get them right.

While for some kind of issues we have counters, tracepoints, or metrics
with a stable API which can reliably indicate the issue, in order to
react to production issues quickly we need to work with the interface
which most kernel developers use when developing: printk. Most
production issues come from unexpected phenomena, and as such usually
the code in question doesn't have easily usable tracepoints or other
counters available for the specific problem being mitigated. We have a
number of lines of monitoring defence against problems in production
(host metrics, process metrics, service metrics, etc), and where it's
not possible to monitor at another level, this kind of netconsole
monitoring is essential.

As you'd expect, monitoring using printk is rather brittle for a number
of reasons -- most notably that the message might disappear entirely in
a new version of the kernel, or that the message may change in some way
that the regex or other classification methods start to silently fail.
Some messages are also particularly unwieldy to detect with such
methods.  This class of issue has happened on a number of occasions,
causing pathologically unhealthy machines with hardware issues to remain
in production and cause extended pain, or be misdiagnosed. As a recent
example, some monitoring around blk_update_request fell out of date and
caused semi-broken machines to remain in production for longer than
would be desirable and cause some churn.

One factor that makes this even harder is that, under normal operation,
many of these messages are never expected to be hit. For example, there
may be some rare hardware bug which you want to detect if it was to ever
happen again, but you don't expect it to. This precludes using something
like a dead detector to determine whether the query is still relevant or
not.

Searching through the codebase to find the message is also extremely
fragile, because many of the messages are further constructed beyond
their callsite (eg. btrfs_printk and other module-specific wrappers,
each with their own functionality). Even if they aren't, guessing the
format and formulation of the underlying message based on the aesthetics
of the message emitted is not a recipe for success at scale, and our
previous issues with fleetwide machine health checking prove that.

This patch provides a solution to the issue of silently changed or
deleted printks: we record all printk format strings known at compile
time into a new .printk_fmts section, both in vmlinux and modules. At
runtime, this can then be iterated by looking at /proc/printk_formats,
which takes the same format as `printk` itself, which we already export
elsewhere (eg. in netconsole):

    # Format: <facility><level><format>\0
    $ perl -p -e 's/\n/\\n/g;s/\0/\n/g' /proc/printk_formats | shuf -n 5
    6                       (parent %llu not aligned to sectorsize %u)\n
    6printk: log_buf_len min size: %d bytes\n
    6pci %04x:%02x:%02x.%d: not ready after %dms; waiting\n
    3SLUB: Unable to add boot slab alias %s to sysfs\n
    4hpet: Global config: Unknown bits %#x\n

This mitigates the majority of cases where we have a highly-specific
printk which we want to match on, as we can now enumerate and check if
the format changed or the printk callsite disappeared entirely in
userspace tools, or as part of the kernel release process. This allows
us to catch these changes to printks we care about earlier and decide
what to do about it.

--

preview v2:

- Don't malloc for module strings
- Change logic so that pos_to_module_idx can be gated behind MODULES
- Always update pos during modules check in !CONFIG_MODULES to avoid
  harmless error message from seq_file
- No need for KEEP() now that section is actually used
- Simplify pr_* macros by putting more logic in _printk_store_fmt
- Add pr_debug/pr_devel
- General tidying

preview v3:

- Simplify iteration: now everything (including vmlinux) is on
  printk_fmt_list. For that reason, the logic is also not gated to
  CONFIG_MODULES.
- Use single_open_size instead of the full seq_file interface -- we
  don't need such a complicated interface, and can just allocate the
  full buffer up front. printk_fmt_total_size is added to handle that.

Signed-off-by: Chris Down <chris@chrisdown.name>