printk: Migrate log_flags to printk_info_flags
In the olden days, `enum log_flags` was part of `struct log`, hence the
name. `struct log` has since been reworked and now this struct is stored
inside `struct printk_info`, but the name was never updated, which is
somewhat confusing (especially since this is at the record level).

Also migrate to printk_ringbuffer.h, since most of the other metadata
structs are over there as well.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Petr Mladek <pmladek@suse.com>