string_helpers: Escape quoting in escape_special
unescape_special already handles unescaping of blackslashed quote
sequences, and as such it seems reasonable to quote them on the other
side to match.

As an example case where this is practically useful: this is needed to
make printk index quoting unambiguous. Compare before, with quotes not
escaped:

    [root@ktst ~]# grep drivers/pci/pci-stub.c:69 /sys/kernel/debug/printk/index/vmlinux
    <4> drivers/pci/pci-stub.c:69 pci_stub_init "pci-stub: invalid ID string "%s"\n"

...and after this patch:

    [root@ktst ~]# grep drivers/pci/pci-stub.c:69 /sys/kernel/debug/printk/index/vmlinux
    <4> drivers/pci/pci-stub.c:69 pci_stub_init "pci-stub: invalid ID string \"%s\"\n"

A new ESCAPE_PRINTK option is also possible, but it seems reasonable to
use the simplest strategy first, since this is already decoded properly.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>