seq_file: Add seq_escape_printf_format
This escapes all characters which would need to be quoted in a typical
printf format, and will be used as part of printk indexing. For example:

    [root@ktst ~]# grep drivers/pci/pci-stub.c:69 /sys/kernel/debug/printk/index/vmlinux
    <4> drivers/pci/pci-stub.c:69 pci_stub_init "pci-stub: invalid ID string \"%s\"\n"

It's also possible to use seq_escape and manually list all of the
characters at the callsite, but that seems non-ideal considering we
already have ESCAPE_{SPECIAL,SPACE} available for use for this purpose.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Al Viro <viro@zeniv.linux.org.uk>