printk: index: Add support for dev_printk
As documented in the patch adding printk indexing, while for some kinds
of issues we have counters, tracepoints, or metrics with a stable
interface which can reliably indicate the issue, in order to react to
production issues quickly we need to work with the interface which most
kernel developers naturally use when developing: printk.

dev_printk is one of the more common subsystem printk-likes for
matching, since device errors brought about by production changes are
one of the most common things we need to match on. Often these errors
were never expected to actually manifest in reality, so there are
typically no metrics for them at the time the incident happens and we
need to match on them.

As such, add the infrastructure for dev_printk to be indexed in the
printk index. The base kernel's printk index is significantly expanded:

Before:

    [root@ktst ~]# wc -l /sys/kernel/debug/printk/index/vmlinux
    4497 /sys/kernel/debug/printk/index/vmlinux

After:

    [root@ktst ~]# wc -l /sys/kernel/debug/printk/index/vmlinux
    5573 /sys/kernel/debug/printk/index/vmlinux

In terms of implementation, dev_printk is now a macro which wraps
_dev_printk in order to trivially disambiguate them, but it's also
possible to have the macro and function have the same name if preferred.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Petr Mladek <pmladek@suse.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>