drm/amdgpu/UAPI: add new PROFILE IOCTL
This adds a new IOCTL currently used to implement querying
and setting the stable power state for GPU profiling.  The
stable pstates use fixed clocks and disable certain power
features in order to get accurate pipeline profiling.

Currently this is handled via sysfs, and that is still
available, but this makes it easier for applications
to utilize.  Note that the power state is global so
setting it will affect all applications.  There are currently
no checks in place to prevent multiple applications from
using this interface, but it doesn't make sense to do
profiling while you have multiple applications running in the
first place, so it's up to the user to ensure this in order
to get good results.

This patch add an interface to query what profiling mode is
currently active and to set enable a profiling mode.

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>