printk: Replace kernel.printk sysctl with more granular controls
The kernel.printk sysctl is unhelpfully overloaded and crusty -- for
example, default_console_loglevel isn't even referenced in the code --
and having it on top of this new per-console loglevel interface will
just make things less intuitive.

Split kernel.printk out into finer-grained sysctls that are easier to
reason about, and warn users that kernel.printk will be removed in a
future version.

Signed-off-by: Chris Down <chris@chrisdown.name>