printk: Introduce new "console" class to expose consoles to sysfs
This patch embeds a class device in the console struct, and registers
each console using it so we can expose attributes in sysfs.

The lifecycle of this classdev looks like this on registration:

    register_console(con)/printk_late_init()
      console_register_device(con)
	con->classdev->flags |= CON_CLASSDEV_ACTIVE
	device_initialize(con->classdev) # refcount++
	device_add(con->classdev)        # refcount++

At stable state, the refcount is two, with further incrementation when
sysfs files are in use.

Destruction looks like this:

    [refcount drops to 0]
	console_classdev_release(con->classdev)
	    con->classdev->flags &= ~CON_CLASSDEV_ACTIVE

    unregister_console(con)
	device_unregister(con->classdev)
	    device_del(con->classdev) # refcount--
	    put_device(con->classdev) # refcount--
	[spin on CON_CLASSDEV_ACTIVE before returning]

Signed-off-by: Chris Down <chris@chrisdown.name>