parisc: fpu: Do not try to index FPU debug printk()s
Commit 337015573718 ("printk: Userspace format indexing support") adds a
struct containing a "fmt" field when CONFIG_PRINTK_INDEX=y.
Unfortunately parisc's FPU code also defines a macro named fmt, which
unhelpfully results in:

    ../arch/parisc/math-emu/decode_exc.c: In function 'decode_fpu':
    ../arch/parisc/math-emu/decode_exc.c:49:25: error: expected identifier before numeric constant
       49 | #define fmt             11      /* bits 19 & 20 */
          |                         ^~
    ../include/linux/printk.h:391:34: note: in expansion of macro 'fmt'
      391 |                                 .fmt = __builtin_constant_p(_fmt) ? (_fmt) : NULL, \

Nobody is likely to actually use these messages and the parisc code is
super old. As such, the simplest and least disruptive thing to do is to
disable printk indexing for these messages, which avoids the struct
appearing in contexts where this fmt macro exists.

Signed-off-by: Chris Down <chris@chrisdown.name>
Cc: Petr Mladek <pmladek@suse.com>
Cc: linux-parisc@vger.kernel.org