printk: console: Create console= parser that supports named options
This will be used in the next patch, and for future changes like the
proposed sync/nothread named options. This avoids having to use sigils
like "/" to get different semantic meaning for different parts of the
option string, and helps to make things more human readable and
easily extensible.

There are no functional changes for existing console= users.

Signed-off-by: Chris Down <chris@chrisdown.name>